--- packages/trpc/server/routers/viewer/organizations/verifyCode.schema.ts
@@ -1,8 +1,0 @@
-import { z } from "zod";
-
-export const ZVerifyCodeInputSchema = z.object({
-  email: z.string().email(),
-  code: z.string(),
-});
-
-export type ZVerifyCodeInputSchema = z.infer<typeof ZVerifyCodeInputSchema>;
